requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
instaloader==4.10
requests==2.31.0

joblib==1.3.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.6.1
xgboost==2.0.3
lightgbm==4.3.0

pillow==10.3.0
opencv-python-headless==4.9.0.80

tensorflow-cpu==2.15.0
mediapipe==0.10.11
